Rename event to point in utils
